Implement fake DNS
This commit implements fake DNS.

Fake DNS implements a UDP listener DNS A record queries on port 53. It
replies with an unused IP address from an address pool, 198.18.0.0/15 by
default. When obtaining a new address from the pool, tun2socks needs to
memorize which name the address belongs to, so that when a client
connects to the address, it can instruct the proxy to connect to the
FQDN. To implement this IP to name mapping, the FakeIP module from clash
is used.